refactor: move database-related files to the database package
