+++ app/src/main/java/campus/tech/kakao/map/database/HistoryDBHelper.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.database
 
 import android.content.Context
 import android.database.sqlite.SQLiteDatabase
